HealthCareforReal/API/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
datasets>=2.14.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Optional dependencies
accelerate>=0.24.0
bitsandbytes>=0.41.0
flash-attn>=2.3.0

# Utility dependencies
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0